requirements.txt
flask>=3.0
openai>=1.0
python-dotenv>=1.0
gunicorn>=21.0
boto3>=1.34.0
dashscope>=1.20.0
pydub>=0.25.0
PyPDF2>=3.0.0
psycopg2-binary>=2.9.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
pypdf>=3.17.0
python-docx>=0.8.11
tiktoken>=0.5.0
Pillow>=10.0.0
